feat: use world write fn for printing.
This allows capturing of stdout.

refactor: restructure parser code
