+++ compose-library/src/foundations/str.rs
@@ -1,10 +1,24 @@
+use std::fmt;
 use compose_library::Value;
 use compose_macros::{cast, ty};
 use ecow::EcoString;
+use compose_library::repr::Repr;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 #[ty(cast, title = "String")]
 pub struct Str(pub EcoString);
+
+impl Repr for Str {
+    fn repr(&self) -> EcoString {
+        self.0.clone()
+    }
+}
+
+impl fmt::Display for Str {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
 
 impl From<EcoString> for Str {
     fn from(s: EcoString) -> Self {
